cli mode integration
