+++ feeluown/components/provider.py
@@ -15,40 +15,42 @@ from PyQt5.QtWidgets import (
 )
 
 
-class LibraryModel(object):
-    def __init__(self, provider, search=None, on_click=None,
-                 icon='♬ ', desc='', **kwargs):
-        self.identifier = provider.identifier
-        self.name = provider.name
+class ProviderModel(object):
+    """
+    一个 ProviderModel 对应左边栏「我的音乐」下的一个项目
+
+    FIXME: 或许不应该命名为 ProviderModel，而是
+    """
+    def __init__(self, name, icon='♬ ', desc='', on_click=None, **kwargs):
+        self.name = name
         self.icon = icon
         self.desc = desc
-        self.search = search
         self._on_click = on_click
 
 
-class LibrariesModel(QAbstractListModel):
-    def __init__(self, libraries=None, parent=None):
+class ProvidersModel(QAbstractListModel):
+    def __init__(self, parent=None):
         super().__init__(parent)
-        self._libraries = libraries or []
 
-    def search(self, keyword):
-        songs = []
-        for library in self._libraries:
-            if not library.search:
-                continue
-            result = library.search(keyword=keyword)
-            _songs = list(result.songs[:20])
-            songs.extend(_songs)
-        return songs
+        self._provider_list = []
+        self._association = {}
 
-    def register(self, library):
-        self._libraries.append(library)
+    def assoc(self, provider_id, pm):
+        self._association[provider_id] = pm
+        self._provider_list.append(provider_id)
+        self.insertRow(len(self._provider_list))
 
-    def __getitem__(self, index):
-        return self._libraries[index]
+    def get(self, provider_id):
+        return self._association.get(provider_id)
+
+    def remove(self, provider_id):
+        if not self._association.pop(provider_id, None):
+            row = self._provider_list.index(provider_id)
+            self._provider_list.remove(provider_id)
+            self.removeRow(row)
 
     def rowCount(self, parent=QModelIndex()):
-        return len(self._libraries)
+        return len(self._provider_list)
 
     def flags(self, index):
         return Qt.ItemIsSelectable | Qt.ItemIsEnabled
@@ -57,20 +59,21 @@ class LibrariesModel(QAbstractListModel):
         if not index.isValid():
             return QVariant()
         row = index.row()
-        if row >= len(self._libraries) or row < 0:
+        if row >= len(self._provider_list) or row < 0:
             return QVariant()
 
-        library = self._libraries[row]
+        provider_id = self._provider_list[row]
+        provider = self._association[provider_id]
         if role == Qt.DisplayRole:
-            return library.icon + ' ' + library.name
+            return provider.icon + ' ' + provider.name
         elif role == Qt.ToolTipRole:
-            return library.desc
+            return provider.desc
         elif role == Qt.UserRole:
-            return library
+            return provider
         return QVariant()
 
 
-class LibrariesView(QListView):
+class ProvidersView(QListView):
     def __init__(self, parent):
         super().__init__(parent)
 
@@ -83,9 +86,9 @@ class LibrariesView(QListView):
         self.clicked.connect(self._on_clicked)
 
     def _on_clicked(self, index):
-        library = index.data(role=Qt.UserRole)
-        if library._on_click:
-            library._on_click()
+        provider = index.data(role=Qt.UserRole)
+        if provider._on_click:
+            provider._on_click()
 
     def sizeHint(self):
         height = 10
