+++ feeluown/plugins/local/__init__.py
@@ -4,7 +4,9 @@ import logging
 from functools import partial
 
 from fuocore.local.provider import LocalProvider
-from feeluown.components.library import LibraryModel
+
+from feeluown.app import App
+from feeluown.components.provider import ProviderModel
 
 __alias__ = '本地音乐'
 __feeluown_version__ = '1.1.0'
@@ -16,15 +18,18 @@ logger = logging.getLogger(__name__)
 
 def enable(app):
     provider = LocalProvider()
-    library = LibraryModel(
-        provider=provider,
-        icon='♪ ',
-        on_click=partial(app.ui.table_container.show_songs, provider.songs),
-        search=None,
-        desc='点击显示所有本地音乐'
-    )
-    app.libraries.register(library)
-    app.ui.table_container.show_songs(provider.songs)
+    logger.info('Register provider: %s' % provider)
+    app.library.register(provider)
+    if app.mode & App.GUIMode:
+        pm = ProviderModel(
+            name='本地音乐',
+            icon='♪ ',
+            desc='点击显示所有本地音乐',
+            on_click=partial(app.ui.table_container.show_songs, provider.songs),
+        )
+        logger.info('Associate %s with %s' % (provider, pm.name))
+        app.providers.assoc(provider.identifier, pm)
+        app.ui.table_container.show_songs(provider.songs)
 
 
 def disable(app):
