+++ feeluown/plugins/neteasemusic/__init__.py
@@ -2,9 +2,13 @@
 
 import logging
 
+from fuocore.netease.provider import provider
+from feeluown.app import App
+from feeluown.components.provider import ProviderModel
+
 from .nem import Nem
 from .consts import LOG_FILE
-from feeluown.config import config
+
 
 __alias__ = '网易云音乐'
 __feeluown_version__ = '1.0.4.2'
@@ -15,10 +19,19 @@ logger = logging.getLogger(__name__)
 
 
 def enable(app):
-    nem = Nem(app)
-    nem.initialize()
-    logger.info('neteasemusic plugin enabled')
+    app.library.register(provider)
+
+    if app.mode & App.GUIMode:
+        nem = Nem(app)
+        pm = ProviderModel(
+            name='网易云音乐',
+            desc='点击可以登录',
+            on_click=nem.ready_to_login,
+        )
+        nem._pm = pm
+        app.providers.assoc(provider.identifier, pm)
 
 
 def disable(app):
-    logger.info('neteasemusic plugin disabled')
+    app.provider.remove(provider.identifier)
+    app.library.deregister(provider)
