+++ feeluown/plugins/neteasemusic/nem.py
@@ -5,7 +5,6 @@ from PyQt5.QtCore import QObject
 
 from fuocore.netease.provider import provider
 from fuocore.netease.models import NUserModel, search
-from feeluown.components.library import LibraryModel
 
 from .login_controller import LoginController
 from .ui import LoginDialog
@@ -25,16 +24,7 @@ class Nem(QObject):
             create_user=LoginController.create,
         )
         self._user = None
-        self._library = LibraryModel(
-            provider=provider,
-            on_click=self.ready_to_login,
-            user=self._user,
-            search=search,
-            desc='点击可以登录'
-        )
-
-    def initialize(self):
-        self._app.libraries.register(self._library)
+        self._pm = None
 
     def ready_to_login(self):
         if self._user is not None:
@@ -58,7 +48,7 @@ class Nem(QObject):
         LoginController.save(user)
         left_panel = self._app.ui.left_panel
         loop = asyncio.get_event_loop()
-        self._library.name = '网易云音乐 - {}'.format(user.name)
+        self._pm.name = '网易云音乐 - {}'.format(user.name)
         playlists = await loop.run_in_executor(None, lambda: user.playlists)
         self._app.playlists.add(playlists)
         self._app.playlists.add(user.fav_playlists, is_fav=True)
