+++ feeluown/plugins/qqmusic/__init__.py
@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
 
 import logging
-from functools import partial
 
 from fuocore.qqmusic.provider import provider
-from fuocore.qqmusic.models import search
-from feeluown.components.library import LibraryModel
+from feeluown.app import App
+from feeluown.components.provider import ProviderModel
 import fuocore.qqmusic.models  # noqa
 
 __alias__ = 'QQ 音乐'
@@ -17,15 +16,17 @@ logger = logging.getLogger(__name__)
 
 
 def enable(app):
-    library = LibraryModel(
-        provider=provider,
-        icon='♫ ',
-        on_click=None,
-        search=search,
-        desc='点击登录 QQ 音乐（未实现，欢迎 PR）',
-    )
-    app.libraries.register(library)
+    app.library.register(provider)
+    if app.mode & App.GUIMode:
+        pm = ProviderModel(
+            name='QQ 音乐',
+            icon='♫ ',
+            desc='点击登录 QQ 音乐（未实现，欢迎 PR）',
+            on_click=None,
+        )
+        app.providers.assoc(provider.identifier, pm)
 
 
 def disable(app):
-    pass
+    app.providers.remove(provider.identifier)
+    app.library.deregister(provider)
